backend-django/requirements.txt
django==4.2.16
djangorestframework==3.15.1
django-cors-headers==4.4.0
python-decouple==3.8
openpyxl==3.1.2
pandas==2.2.2
python-dotenv==1.0.1
whitenoise==6.8.0